backend/requirements.txt
fastapi
uvicorn
pydantic
pyjwt==2.8.0
